refactor(teams): move PERM_META beside the permission bits
